Create SteerControl
Currently, this code is a copy of LeftRightDoubleDrive, this code will be used as a base for the new SteerControl program. The goals of the SteerControl program are as follows, have the left stick drive the robot forwards and backwards, have the right stick steer by subtracting power from one side and adding it to the other depending on how far the right stick is pushed with the right stick being pushed all the way in one direction causing the robot to spin in place.

Develop code to operate the robotic arm, an expansion hub may be added or two servos will operate the base or the arm, while one servo will operate the claw.

Something to start preparing for is strafing, we need to research this, the controls would be the left and right bumpers on the controller.

As a bonus, we can also develop code to get the robot to drive straight when instructed to wether by autonomous or TeleOp, as sometimes the robot does not drive straight even when programmed or directed to do so.